// js-client/src/client/models/Page_StaffRolesSchema_.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { StaffRolesSchema } from './StaffRolesSchema';

export type Page_StaffRolesSchema_ = {
    items: Array<StaffRolesSchema>;
    total?: number;
    page?: number;
    size?: number;
    pages?: number;
};