// js-client/src/client/models/Page_UserOut_.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { UserOut } from './UserOut';

export type Page_UserOut_ = {
    items: Array<UserOut>;
    total?: number;
    page?: number;
    size?: number;
    pages?: number;
};